requirements.txt
python-dotenv~=1.0.0

requests~=2.31.0
aiogram~=2.25.1
geopy~=2.3.0
loguru~=0.7.0
peewee~=3.15.4